abstracted wfs getGranules to use convert/granules. Began converting STAC search endpoint.
