+++ search/lib/api/wfs.js
@@ -1,9 +1,7 @@
 const express = require('express');
 const { wfs, generateAppUrl, logger } = require('../util');
 const cmr = require('../cmr');
-const settings = require('../settings');
 const convert = require('../convert');
-const { generateAppUrlWithoutRelativeRoot, generateSelfUrl, extractParam } = require('../util');
 
 async function getCollections (request, response) {
   logger.info(`GET ${request.params.providerId}/collections`);
@@ -37,44 +35,44 @@ async function getCollection (request, response) {
 }
 
 async function getGranules (request, response) {
-  logger.info(`GET /${request.params.providerId}/collections/${request.params.collectionId}/items`);
-  const event = request.apiGateway.event;
   const conceptId = request.params.collectionId;
+  logger.info(`GET /${request.params.providerId}/collections/${conceptId}/items`);
+  const event = request.apiGateway.event;
   const params = Object.assign({ collection_concept_id: conceptId }, cmr.convertParams(cmr.WFS_PARAMS_CONVERSION_MAP, request.query));
   const granules = await cmr.findGranules(params);
-
-  const currPage = parseInt(extractParam(event.queryStringParameters, 'page_num', '1'), 10);
-  const nextPage = currPage + 1;
-  const prevPage = currPage - 1;
-  const newParams = { ...event.queryStringParameters } || {};
-  newParams.page_num = nextPage;
-  const newPrevParams = { ...event.queryStringParameters } || {};
-  newPrevParams.page_num = prevPage;
-  const prevResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newPrevParams);
-  const nextResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newParams);
-
-  const granulesResponse = {
-    type: 'FeatureCollection',
-    stac_version: settings.stac.version,
-    features: granules.map(gran => convert.cmrGranToFeatureGeoJSON(event, gran)),
-    links: [
-      {
-        rel: 'self',
-        href: generateSelfUrl(event)
-      },
-      {
-        rel: 'next',
-        href: nextResultsLink
-      }
-    ]
-  };
-
-  if (currPage > 1 && granulesResponse.links.length > 1) {
-    granulesResponse.links.splice(1, 0, {
-      rel: 'prev',
-      href: prevResultsLink
-    });
-  }
+  const granulesResponse = convert.cmrGranulesToFeatureCollection(event, granules);
+  // const currPage = parseInt(extractParam(event.queryStringParameters, 'page_num', '1'), 10);
+  // const nextPage = currPage + 1;
+  // const prevPage = currPage - 1;
+  // const newParams = { ...event.queryStringParameters } || {};
+  // newParams.page_num = nextPage;
+  // const newPrevParams = { ...event.queryStringParameters } || {};
+  // newPrevParams.page_num = prevPage;
+  // const prevResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newPrevParams);
+  // const nextResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newParams);
+  //
+  // const granulesResponse = {
+  //   type: 'FeatureCollection',
+  //   stac_version: settings.stac.version,
+  //   features: granules.map(gran => convert.cmrGranToFeatureGeoJSON(event, gran)),
+  //   links: [
+  //     {
+  //       rel: 'self',
+  //       href: generateSelfUrl(event)
+  //     },
+  //     {
+  //       rel: 'next',
+  //       href: nextResultsLink
+  //     }
+  //   ]
+  // };
+  //
+  // if (currPage > 1 && granulesResponse.links.length > 1) {
+  //   granulesResponse.links.splice(1, 0, {
+  //     rel: 'prev',
+  //     href: prevResultsLink
+  //   });
+  // }
 
   response.status(200).json(granulesResponse);
 }
