+++ search/lib/convert/granules.js
@@ -176,46 +176,40 @@ function cmrGranToFeatureGeoJSON (event, cmrGran) {
 }
 
 function cmrGranulesToFeatureCollection (event, cmrGrans) {
-  if (event.queryStringParameters.page_num > 1) {
-    const currPage = event.queryStringParameters.page_num;
-    const nextPage = currPage + 1;
-    const prevPage = currPage - 1;
-    const newParams = { ...event.queryStringParameters } || {};
-    newParams.page_num = nextPage;
-    const newPrevParams = { ...event.queryStringParameters } || {};
-    newPrevParams.page_num = prevPage;
-    const prevResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newPrevParams);
-    const nextResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newParams);
-
-    return {
-      type: 'FeatureCollection',
-      stac_version: settings.stac.version,
-      features: cmrGrans.map(g => cmrGranToFeatureGeoJSON(event, g)),
-      links: {
-        self: generateSelfUrl(event),
-        prev: prevResultsLink,
-        next: nextResultsLink
-      }
-    };
-  }
-
   const currPage = parseInt(extractParam(event.queryStringParameters, 'page_num', '1'), 10);
   const nextPage = currPage + 1;
+  const prevPage = currPage - 1;
   const newParams = { ...event.queryStringParameters } || {};
   newParams.page_num = nextPage;
+  const newPrevParams = { ...event.queryStringParameters } || {};
+  newPrevParams.page_num = prevPage;
+  const prevResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newPrevParams);
   const nextResultsLink = generateAppUrlWithoutRelativeRoot(event, event.path, newParams);
 
-  return {
+  const granulesResponse = {
     type: 'FeatureCollection',
     stac_version: settings.stac.version,
     features: cmrGrans.map(g => cmrGranToFeatureGeoJSON(event, g)),
     links: [
       {
-        self: generateSelfUrl(event),
-        next: nextResultsLink
+        rel: 'self',
+        href: generateSelfUrl(event)
+      },
+      {
+        rel: 'next',
+        href: nextResultsLink
       }
     ]
   };
+
+  if (currPage > 1 && granulesResponse.links.length > 1) {
+    granulesResponse.links.splice(1, 0, {
+      rel: 'prev',
+      href: prevResultsLink
+    });
+  }
+
+  return granulesResponse;
 }
 
 module.exports = {
