+++ search/lib/settings.js
@@ -12,7 +12,7 @@ function getLoggerSettings () {
 function getStacSettings () {
   const stacSettings = {};
 
-  stacSettings.version = process.env.STAC_VERSION || '0.8.0';
+  stacSettings.version = process.env.STAC_VERSION || '1.0.0-beta.1';
   stacSettings.stacRelativePath = process.env.STAC_RELATIVE_PATH || '/stac';
 
   return stacSettings;
