+++ search/lib/validator/stacValidator.js
@@ -1,9 +1,12 @@
 
-function validateStac (stacItem) {
-  if (!stacItem) throw new Error('Missing stacItem');
+function validateStac (featureCollection) {
+  if (!featureCollection) throw new Error('Missing Feature Collection');
   try {
-    if (!stacItem.stac_version || !stacItem.id || !stacItem.description || !stacItem.links) {
-      throw Error('Missing required fields');
+    for (const stacItem of featureCollection.features) {
+      if (!stacItem.stac_version || !stacItem.id || !stacItem.links) {
+        console.error('Missing required fields');
+        throw Error('Missing required fields');
+      }
     }
   } catch (e) {
     return false;
