+++ search/serverless-offline.yml
@@ -19,7 +19,7 @@ functions:
     environment:
       LOG_LEVEL: info
       LOG_DISABLED: false
-      STAC_VERSION: 0.8.0
+      STAC_VERSION: 1.0.0-beta.1
       CMR_STAC_RELATIVE_ROOT_URL: /cmr-stac
 
 plugins:
